Generate hooks in a single to_nice_yaml invocation
Invoking to_nice_yaml on a scalar value generates a two-line YAML document "value\n...\n" which corrupts the value of a hook such as healthcheck which is just a single string URL. Instead of looping over each of the hooks, define a dictionary with hooks at the top level and have to_nice_yaml transform it in a single invocation. Fixes #117.